Move files to folder and fix index file with router
