+++ src/index.js
@@ -1,17 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import App from './Components/App/App';
+import { BrowserRouter } from 'react-router-dom';
 import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+const router = (<BrowserRouter> <App /> </BrowserRouter>);
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+ReactDOM.render(router, document.getElementById('root'));
